+++ source/Module/Rule.SqlProtocol/SqlProtocolRule.psm1
@@ -1,0 +1,63 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+using module .\..\Common\Common.psm1
+using module .\..\Rule\Rule.psm1
+#header
+
+<#
+    .SYNOPSIS
+        SQL Server protocol rule
+    .DESCRIPTION
+        The SqlProtocolRule class is used to manage SQL authentication logins
+    .PARAMETER ProtocolName
+        The SQL Server protocol name
+    .PARAMETER Enabled
+        The SQL Server protocol status
+#>
+class SqlProtocolRule : Rule
+{
+    [string] $ProtocolName
+    [string] $Enabled <#(ExceptionValue)#>
+
+    <#
+        .SYNOPSIS
+            Default constructor to support the AsRule cast method
+    #>
+    SqlProtocolRule ()
+    {
+    }
+
+    <#
+        .SYNOPSIS
+            Used to load PowerSTIG data from the processed data directory
+        .PARAMETER Rule
+            The STIG rule to load
+    #>
+    SqlProtocolRule ([xml.xmlelement] $Rule) : base ($Rule)
+    {
+    }
+
+    <#
+        .SYNOPSIS
+            The Convert child class constructor
+        .PARAMETER Rule
+            The STIG rule to convert
+        .PARAMETER Convert
+            A simple bool flag to create a unique constructor signature
+    #>
+    SqlProtocolRule ([xml.xmlelement] $Rule, [switch] $Convert) : base ($Rule, $Convert)
+    {
+    }
+
+    <#
+        .SYNOPSIS
+            Creates class specifc help content
+    #>
+    [PSObject] GetExceptionHelp()
+    {
+        return @{
+            Value = "15"
+            Notes = $null
+        }
+    }
+}